# src/main/resources/openapi/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8082
  description: Generated server url
tags:
- name: Account API
  description: Gestión de cuentas bancarias
paths:
  /v1.0/accounts/{id}:
    get:
      tags:
      - Account API
      summary: Obtener una cuenta por id
      description: Busca una cuenta bancaria por id
      operationId: getAccountById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Cuenta no encontrada
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
        "500":
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
        "200":
          description: Cuenta encontrada
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
    put:
      tags:
      - Account API
      summary: Actualizar saldo de una cuenta
      description: Modifica el saldo de una cuenta bancaria
      operationId: updateBalance
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: number
              format: double
        required: true
      responses:
        "404":
          description: Cuenta no encontrada
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
        "500":
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
        "200":
          description: Saldo actualizado correctamente
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
    delete:
      tags:
      - Account API
      summary: Eliminar una cuenta bancaria
      description: Elimina una cuenta bancaria del sistema
      operationId: deleteAccount
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Cuenta no encontrada
        "200":
          description: Cuenta eliminada exitosamente
        "500":
          description: Error interno del servidor
  /v1.0/accounts:
    get:
      tags:
      - Account API
      summary: Obtener todas las cuentas
      description: Lista todas las cuentas bancarias
      operationId: getAllAccounts
      responses:
        "500":
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        "200":
          description: Cuentas obtenidas correctamente
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
    post:
      tags:
      - Account API
      summary: Crear una cuenta bancaria
      description: Registra una nueva cuenta en el sistema
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
        required: true
      responses:
        "201":
          description: Cuenta creada exitosamente
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
        "500":
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
        "400":
          description: Error en la validación de la cuenta
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
        "404":
          description: Cliente no encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
  /v1.0/accounts/byNumber/{number}:
    get:
      tags:
      - Account API
      summary: Obtener una cuenta por número
      description: Busca una cuenta bancaria por su número único
      operationId: getAccountByNumber
      parameters:
      - name: number
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Cuenta no encontrada
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
        "500":
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
        "200":
          description: Cuenta encontrada
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
components:
  schemas:
    Account:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        type:
          type: string
        customerId:
          type: string
        balance:
          type: number
          format: double
        hasMaintenanceFee:
          type: boolean
        transactionLimit:
          type: integer
          format: int32
        authorizedSigners:
          type: array
          items:
            type: string